+++ src/pytest_forked/__init__.py
@@ -44,13 +44,14 @@ def pytest_load_initial_conftests(early_config, parser, args):
 
 
 @pytest.hookimpl(tryfirst=True)
-def pytest_runtest_protocol(item):
+def pytest_runtest_protocol(item, nextitem):
     if item.config.getvalue("forked") or item.get_closest_marker("forked"):
         ihook = item.ihook
         ihook.pytest_runtest_logstart(nodeid=item.nodeid, location=item.location)
         reports = forked_run_report(item)
         for rep in reports:
             ihook.pytest_runtest_logreport(report=rep)
+        item.session._setupstate.teardown_exact(nextitem)
         ihook.pytest_runtest_logfinish(nodeid=item.nodeid, location=item.location)
         return True